virtual prototype
This commit is a "proof of concept" of sorts to model possible ways of
using beamforming with radio waves to produce a signal powerful enough
to remove adverse bacteria, pollution and other contaminates from large
amounts of water in order to make it fit for consumption by all living
organisms. Who knows? It might even lead to something as wild as
interdimensional/time travel. Then again, I'm just thinking about the
water right now. Any help in the formula department would be greatly
appreciated. I think I have a solid grasp of the physics, however,
having not received much formal training beyond quadratic equations, I
could use a little help. Specifically a formula to determine which
frequencies to use given the ambient conditions at the respective
transmit/receive stations and if there is a way to transmit this
information betweeen the transmiter/receiver prior to an attempted
connection. Other than that, It should be ezy peazy. I've already been
able to transmit via the beamforming method on a $35 Baofeng dual
frequency HAM. Scaled up, this should be pretty impressive.